feat: add text-to-speech feature with Kokoro TTS integration
