+++ DesktopEye.Common/Application/ViewModels/AudioPlayerViewModel.cs
@@ -3,8 +3,8 @@ using System.IO;
 using System.Threading.Tasks;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
+using DesktopEye.Common.Domain.Features.TextToSpeech;
 using DesktopEye.Common.Domain.Models;
-using DesktopEye.Common.Infrastructure.Services.TTS;
 using SoundFlow.Components;
 using SoundFlow.Providers;
 using SoundFlow.Backends.MiniAudio;
