+++ DesktopEye.Common/Domain/Models/TextToSpeech/TtsType.cs
@@ -1,0 +1,6 @@
+﻿namespace DesktopEye.Common.Domain.Models.TextToSpeech;
+
+public enum TtsType
+{
+    KokoroTts,
+}
